[hipTensor] hipTensor buildable on Windows

+++ projects/hiptensor/library/src/contraction/hiptensor_contraction.cpp
@@ -409,7 +409,7 @@ hiptensorStatus_t hiptensorContract(const hiptensorHandle_t handle,
             // log perf metrics (not name/id)
             snprintf(msg,
                      sizeof(msg),
-                     "KernelId: %lu KernelName: %s, %0.3f ms, %0.3f TFlops/s, %0.3f GB/s",
+                     "KernelId: %zu KernelName: %s, %0.3f ms, %0.3f TFlops/s, %0.3f GB/s",
                      metrics.mKernelUid,
                      metrics.mKernelName.c_str(),
                      metrics.mAvgTimeMs,
@@ -449,9 +449,9 @@ hiptensorStatus_t hiptensorContract(const hiptensorHandle_t handle,
     {
         snprintf(msg,
                  sizeof(msg),
-                 "Insufficient workspace: req: %lu alloc: %lu (%s)",
+                 "Insufficient workspace: req: %zu alloc: %llu (%s)",
                  cSolution->workspaceSize(),
-                 workspaceSize,
+                 static_cast<unsigned long long>(workspaceSize),
                  hiptensorGetErrorString(errorCode));
         logger->logError("hiptensorContraction", msg);
     }
@@ -750,7 +750,7 @@ hiptensorStatus_t contractionInitPlan(const hiptensorHandle_t              handl
     // Log the selected contraction solution and selection timing
     snprintf(msg,
              sizeof(msg),
-             "Algo: %d, KernelId: %lu, KernelName: %s, SelectionTime: %0.3f ms",
+             "Algo: %d, KernelId: %zu, KernelName: %s, SelectionTime: %0.3f ms",
              pref->mSelectionAlgorithm,
              winner->uid(),
              winner->kernelName().c_str(),
